Move JavaVersion into version_resolver::version_metadata
