+++ src/app/generator.rs
@@ -9,8 +9,7 @@ use bytes::Bytes;
 use futures::future::join_all;
 use futures::{join, FutureExt};
 use miette::{IntoDiagnostic, Result};
-use version_resolver::minecraft::JavaVersion;
-use version_resolver::version_metadata::MinecraftVersionList;
+use version_resolver::version_metadata::{JavaVersion, MinecraftVersionList};
 use std::future::Future;
 use std::pin::Pin;
 use std::sync::Arc;
